Bugfix TTL() (#33)
* update Makefile

* update Makefile

* update Makefile

* support scan

* optimize code

* fix if the key already expired

* optimize code

* optimize code

* bugfix scan()

* Fix conflict

* Add benchmark

* Add more tests

* bugfix benchmark

* fix conflict

* Optimize code

* Complete keys interfaces

* Update code comment

* Bugfix TTL()

* bugfix TTL()

* Fix conflicts

* remove executable file

+++ include/blackwidow/blackwidow.h
@@ -433,7 +433,7 @@ class BlackWidow {
   // return -2 if the key does not exist
   // return -1 if the key exists but has not associated expire
   // return > 0 TTL in seconds
-  std::map<DataType, int32_t> TTL(const Slice& key,
+  std::map<DataType, int64_t> TTL(const Slice& key,
               std::map<DataType, Status>* type_status);
 
  private:
