+++ src/redis.h
@@ -51,8 +51,7 @@ class Redis {
   virtual Status Expireat(const Slice& key,
                           int32_t timestamp) = 0;
   virtual Status Persist(const Slice& key) = 0;
-  virtual Status TTL(const Slice& key,
-                     int32_t* timestamp) = 0;
+  virtual Status TTL(const Slice& key, int64_t* timestamp) = 0;
 
  protected:
   LockMgr* lock_mgr_;
