+++ src/redis_hashes.h
@@ -60,7 +60,7 @@ class RedisHashes : public Redis {
                     int64_t* count, std::string* next_key) override;
   virtual Status Expireat(const Slice& key, int32_t timestamp) override;
   virtual Status Persist(const Slice& key) override;
-  virtual Status TTL(const Slice& key, int32_t* timestamp) override;
+  virtual Status TTL(const Slice& key, int64_t* timestamp) override;
 
  private:
   std::vector<rocksdb::ColumnFamilyHandle*> handles_;
