+++ src/redis_sets.h
@@ -62,7 +62,7 @@ class RedisSets : public Redis {
                     int64_t* count, std::string* next_key) override;
   virtual Status Expireat(const Slice& key, int32_t timestamp) override;
   virtual Status Persist(const Slice& key) override;
-  virtual Status TTL(const Slice& key, int32_t* timestamp) override;
+  virtual Status TTL(const Slice& key, int64_t* timestamp) override;
 
  private:
   std::vector<rocksdb::ColumnFamilyHandle*> handles_;
