+++ src/redis_strings.h
@@ -66,7 +66,7 @@ class RedisStrings : public Redis {
                     int64_t* count, std::string* next_key) override;
   virtual Status Expireat(const Slice& key, int32_t timestamp) override;
   virtual Status Persist(const Slice& key) override;
-  virtual Status TTL(const Slice& key, int32_t* timestamp) override;
+  virtual Status TTL(const Slice& key, int64_t* timestamp) override;
 };
 
 }  //  namespace blackwidow
